requirements.txt
pdf2image
Pillow
tk